Working with signup
feat: Update the General service to support alerts and toast. Add a new service for data storage. Implement two structures for the user (data sent and received). Implement the use of request and interact with the api to perform the registration.

The registration process is almost complete, the only thing missing is the implementation of the native plugins and other validations.